Build MEX functions with static libs2hat + static libfftw3
This is required to build the MEX functions correctly since Matlab
ships its own (internal) FFTW library. If the dynamic library is used,
the symbols resolve to Matlab's version which may conflict with the
system version that the s2hat library was compiled against.

The only way I've found to get around this is to just statically compile
in the entire libfftw3 library. It's unfortunate this is necessary, but
Matlab. *sigh*