Provide workaround for __FILE__ being full path in debug statements.
The mex compiler expands the source code paths to full paths, and since
compilers often use the command line provided name to fill the __FILE__
macro, this causes all debugging statements to include the full path in
the output.

By explicitly setting the file name at the beginning of the file with
the #line macro, we can get the file name without any extra path
information.